+++ facebook-core/src/main/java/com/facebook/FacebookSdkVersion.kt
@@ -20,5 +20,5 @@
 package com.facebook
 
 internal object FacebookSdkVersion {
-  const val BUILD = "11.3.0"
+  const val BUILD = "12.0.0"
 }